+++ src/pages/deck-import/ui/DeckImportView.stories.tsx
@@ -6,8 +6,8 @@
 
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { SAMPLE_CSV_TEXT } from "@/features/import";
-import type { DeckImportPreview } from "@/features/import";
+import { SAMPLE_CSV_TEXT } from "@/features/deck/import";
+import type { DeckImportPreview } from "@/features/deck/import";
 import { INITIAL_VIEWPORTS } from "@/storybook/storybookViewports";
 
 import { DeckImportView as Template } from "./DeckImportView";
