+++ steiger.config.ts
@@ -16,7 +16,6 @@ export default defineConfig([
     files: [
       "./src/entities/study-progress/**",
       "./src/features/deck-editor/**",
-      "./src/features/export/**",
       "./src/features/settings/**",
     ],
     rules: {
